# Conflicts:
#	beaver-iot-parent/pom.xml
#	core/context/src/main/java/com/milesight/beaveriot/context/security/SecurityUserContext.java
#	core/context/src/main/java/com/milesight/beaveriot/context/util/ExchangeContextHelper.java
#	services/authentication/authentication-service/src/main/java/com/milesight/beaveriot/authentication/provider/CustomOAuth2PasswordAuthenticationConverter.java
#	services/permission/permission-service/src/main/java/com/milesight/beaveriot/permission/aspect/TenantAspect.java
#	services/user/user-api/src/main/java/com/milesight/beaveriot/user/facade/IUserFacade.java
#	services/user/user-service/src/main/java/com/milesight/beaveriot/user/controller/UserController.java

+++ services/user/user-service/src/main/java/com/milesight/beaveriot/user/facade/UserFacade.java
@@ -149,6 +149,9 @@ public class UserFacade implements IUserFacade {
 
     @Override
     public void deleteResource(ResourceType resourceType, List<Long> resourceIds) {
+        if (resourceIds == null || resourceIds.isEmpty()){
+            return;
+        }
         List<RoleResourcePO> roleResourcePOS = roleResourceRepository.findAll(filterable -> filterable.in(RoleResourcePO.Fields.resourceId, resourceIds.toArray()).eq(RoleResourcePO.Fields.resourceType, resourceType));
         if (roleResourcePOS != null && !roleResourcePOS.isEmpty()) {
             roleResourceRepository.deleteAll(roleResourcePOS);
